Organize core classes into functional packages
