+++ core/src/main/java/com/spamalot/arcade/robostar/entity/Bomb.java
@@ -1,8 +1,10 @@
-package com.spamalot.arcade.robostar;
+package com.spamalot.arcade.robostar.entity;
 
 import com.badlogic.gdx.graphics.Color;
 import com.badlogic.gdx.graphics.glutils.ShapeRenderer;
 import com.badlogic.gdx.math.Vector2;
+
+import com.spamalot.arcade.robostar.world.WorldUtils;
 
 public class Bomb {
   private Vector2 pos = new Vector2();
