+++ core/src/main/java/com/spamalot/arcade/robostar/entity/Boss.java
@@ -1,9 +1,11 @@
-package com.spamalot.arcade.robostar;
+package com.spamalot.arcade.robostar.entity;
 
 import com.badlogic.gdx.graphics.Color;
 import com.badlogic.gdx.graphics.glutils.ShapeRenderer;
 import com.badlogic.gdx.math.MathUtils;
 import com.badlogic.gdx.math.Vector2;
+
+import com.spamalot.arcade.robostar.world.WorldUtils;
 
 public class Boss {
   private Vector2 pos = new Vector2();
