+++ core/src/main/java/com/spamalot/arcade/robostar/entity/Enemy.java
@@ -1,10 +1,13 @@
-package com.spamalot.arcade.robostar;
+package com.spamalot.arcade.robostar.entity;
 
 import com.badlogic.gdx.graphics.Color;
 import com.badlogic.gdx.graphics.glutils.ShapeRenderer;
 import com.badlogic.gdx.math.MathUtils;
 import com.badlogic.gdx.math.Vector2;
 import com.badlogic.gdx.utils.Array;
+
+import com.spamalot.arcade.robostar.event.BossBuildProgressEvent;
+import com.spamalot.arcade.robostar.event.EventDispatcher;
 
 public class Enemy {
   public enum Type {
