+++ core/src/main/java/com/spamalot/arcade/robostar/event/EventDispatcher.java
@@ -1,4 +1,4 @@
-package com.spamalot.arcade.robostar;
+package com.spamalot.arcade.robostar.event;
 
 import java.util.function.Consumer;
