+++ core/src/main/java/com/spamalot/arcade/robostar/screen/GameState.java
@@ -1,4 +1,4 @@
-package com.spamalot.arcade.robostar;
+package com.spamalot.arcade.robostar.screen;
 
 public enum GameState {
   MENU, PLAY, GAME_OVER
