+++ core/src/main/java/com/spamalot/arcade/robostar/screen/PlayScreen.java
@@ -1,9 +1,14 @@
-package com.spamalot.arcade.robostar;
+package com.spamalot.arcade.robostar.screen;
 
 import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.Screen;
 import com.badlogic.gdx.graphics.GL20;
 
+import com.spamalot.arcade.robostar.GameRoot;
+import com.spamalot.arcade.robostar.event.BossBuildProgressEvent;
+import com.spamalot.arcade.robostar.event.EventDispatcher;
+import com.spamalot.arcade.robostar.event.SimpleEventDispatcher;
+import com.spamalot.arcade.robostar.world.WorldManager;
 
 /**
  * Main play screen delegating world updates and rendering.
